pyproject.toml
[tool.poetry]
name = "reason-net"
version = "0.1.0"
description = ""
authors = ["samsja <sami.jaghouar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch =  ">=2.1.0"
wandb = ">=0.15.3"
lightning = ">=2.1.0"
jaxtyping = ">=0.2.25"
hydra-core = "^1.3.2"
pydantic = "^2.5.3"
einops = "^0.7.0"
beartype = "^0.16.4"
cyclopts = "^1.2.0"
rich = "^13.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
ruff = "^0.1.9"
pre-commit = "^3.3.2"
jupyterlab = "^4.0.0"
jupyterlab-code-formatter = "^2.2.1"
mypy = "^1.5.1"

[tool.mypy]
ignore_missing_imports = true

[tool.ruff]
ignore = [
    "E501",  # line too long, handled by black
    "F722", # allow jaxtyping see here https://docs.kidger.site/jaxtyping/faq/#flake8-or-ruff-are-throwing-an-error
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
